requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
timm>=0.9.0
matplotlib>=3.7.0
Pillow>=10.0.0
pytorch-msssim>=1.0.0
pandas>=2.0.0
pyvips>=2.2.0
valis-wsi>=1.0.0
